+++ src/index.ts
@@ -3,4 +3,4 @@ export * from "./crudify/crudify.service";
 export * from "./crudify/crudify.decorator";
 export * from "./crudify/interface/crudify.interface";
 
-export * from "./logger/logger.module";
+export * from "./crudify/crudify.module";