+++ src/logger/logger.module.ts
@@ -1,31 +1,12 @@
-import { DynamicModule, Module, OnModuleInit } from "@nestjs/common";
+import { LoggerModuleOptions } from "./logger.interface";
 import * as errsole from "errsole";
-import ErrsoleMongoDB from "errsole-mongodb";
-import { ILoggerModuleOptions } from "./logger.interface";
 
-@Module({})
-export class CrudifyLoggerModule implements OnModuleInit {
-  private static options: ILoggerModuleOptions;
-  static forRoot(options: ILoggerModuleOptions): DynamicModule {
-    this.options = {
-      uri: options.uri || "mongodb://localhost:27017/",
-      dbName: options.dbName || "logs",
-    };
-    return {
-      module: CrudifyLoggerModule,
-    };
-  }
-
-  onModuleInit() {
-    const mongoUri =
-      CrudifyLoggerModule.options?.uri || "mongodb://localhost:27017/";
-    const dbName = CrudifyLoggerModule.options?.dbName || "logs";
-    if (!mongoUri) {
+export class CrudifyLoggerModule {
+  initModule(options: LoggerModuleOptions) {
+    if (options?.disabled) return;
+    if (!options.uri) {
       throw new Error("MongoDB URI not found");
     }
-    errsole.initialize({
-      storage: new ErrsoleMongoDB(mongoUri, dbName!),
-      ...CrudifyLoggerModule.options.options,
-    });
+    errsole.initialize(options.options);
   }
 }