+++ src/shared/const.namespace.ts
@@ -1,0 +1,3 @@
+export namespace CONST {
+  export const MONGO_URI: string = "mongodb://127.0.0.1:27017/";
+}